books-cover/requirements.txt
psycopg2
python-dotenv
langchain-openai
langchain-core